Make library Fixes #6
