+++ .gitignore
@@ -129,3 +129,6 @@ dmypy.json
 .pyre/
 
 .vscode/
+
+# osx 
+.*
