+++ images4web/requirements.txt
@@ -1,0 +1,2 @@
+files==1.1.1
+Pillow==9.2.0
