--- images4web/settings.py
@@ -1,9 +1,0 @@
-from glob import glob
-
-
-global syncDelay
-global compressionQuality
-
-# the delay for the syncronisation in seconds
-syncDelay = 1
-compressionQuality = 20
