+++ images4web/settings/parameters.py
@@ -1,0 +1,9 @@
+from glob import glob
+
+
+global syncDelay
+global compressionQuality
+
+# the delay for the syncronisation in seconds
+syncDelay = 1
+compressionQuality = 20
