+++ images4web/test_syncronize.py
@@ -1,0 +1,6 @@
+from classes.syncronize import *
+import os
+
+inputImagePath = os.path.join(os.getcwd(), 'images4web/images')
+outputImagePath = os.path.join(inputImagePath, '.compressed4web')
+Sync1 = Syncronize(inputImagePath, outputImagePath, compressionQuality=10)
